[Silver I] Title: 스티커, Time: 796 ms, Memory: 111680 KB -BaekjoonHub

+++ 백준/Silver/9465. 스티커/스티커.java
@@ -12,14 +12,14 @@ public class Main {
             StringTokenizer top = new StringTokenizer(br.readLine());
             StringTokenizer bottom = new StringTokenizer(br.readLine());
 
-            long[][] dp = new long[2][n + 1];
-            
-            dp[0][1] = Long.parseLong(top.nextToken());
-            dp[1][1] = Long.parseLong(bottom.nextToken());
+            int[][] dp = new int[2][n + 1];
+
+            dp[0][1] = Integer.parseInt(top.nextToken());
+            dp[1][1] = Integer.parseInt(bottom.nextToken());
 
             for (int j = 2; j <= n; j++) {
-                dp[0][j] = Math.max(dp[1][j - 1], dp[1][j - 2]) + Long.parseLong(top.nextToken());
-                dp[1][j] = Math.max(dp[0][j - 1], dp[0][j - 2]) + Long.parseLong(bottom.nextToken());
+                dp[0][j] = Math.max(dp[1][j - 1], dp[1][j - 2]) + Integer.parseInt(top.nextToken());
+                dp[1][j] = Math.max(dp[0][j - 1], dp[0][j - 2]) + Integer.parseInt(bottom.nextToken());
             }
 
             System.out.println(Math.max(dp[0][n], dp[1][n]));
